fix(frontend): export material type
